C&P&H template

+++ CPH.cpp
@@ -1,7 +1,7 @@
 #include <bits/stdc++.h>
-#define FOR(n) for(ull i=0;i<n;i++)
-#define JFOR(n) for(ull j=0;j<n;j++)
-#define KFOR(n) for(ull k=0;k<n;k++)
+#define FOR(n) for(int i=0;i<n;i++)
+#define JFOR(n) for(int j=0;j<n;j++)
+#define KFOR(n) for(int k=0;k<n;k++)
 #define ioopt ios_base::sync_with_stdio(0);cin.tie(0);
 //Don't use macros below if U R using ioopt!!
 #define RI(n) scanf("%d",&n)
